+++ __test__/fields/main.test.ts
@@ -4,6 +4,13 @@ describe("package.json main field", () => {
 
   const cwd = process.cwd();
 
+  test("should throw on non string main field", () => {
+
+    // @ts-ignore
+    expect(analize(cwd, { main: 100 })).rejects.toThrow(TypeError);
+
+  });
+
   test("should read main field", async () => {
 
     const analized = await analize(cwd, { main: "out/lib.js" });